refactor(verify): originator-profile-set 検証モジュールを分割
verify-ops.ts に詰め込まれていた OP/PA/WMP 個別の検証ロジックと
テストフィクスチャを以下のモジュールに切り出し、認知負荷を下げる。

- op-verifier.ts: CP 以外の OP 単体署名検証
- annotations.ts: annotations プロパティの署名検証と有効期限チェック
- media.ts: media プロパティの署名検証
- helper.ts: OPS テスト用共通フィクスチャ buildOpsFixture

これに合わせて verify-ops.test.ts を annotations.test.ts /
media.test.ts に分割。inspector の dev fixture も将来複数 PA を
配置できるよう _annotations/ ディレクトリ + glob 配信に整理。

機能・公開 API には変更なし。
