--- apps/inspector/dev/src/pages/api/_annotation.json
@@ -1,23 +1,0 @@
-{
-  "@context": [
-    "https://www.w3.org/ns/credentials/v2",
-    "https://originator-profile.org/ns/credentials/v1",
-    "https://originator-profile.org/ns/cip/v1",
-    {
-      "@language": "ja-JP"
-    }
-  ],
-  "type": ["VerifiableCredential", "Certificate"],
-  "issuer": "dns:localhost",
-  "credentialSubject": {
-    "id": "dns:localhost",
-    "type": "CertificateProperties",
-    "description": "Example Certificate",
-    "certificationSystem": {
-      "id": "urn:uuid:de5d6e80-10a5-404f-b4d3-e9f0e6926a21",
-      "type": "CertificationSystem",
-      "name": "Example Certification System",
-      "description": "Example Certification System Description"
-    }
-  }
-}
