+++ apps/inspector/dev/src/pages/api/_annotations/existence.json
@@ -1,0 +1,30 @@
+{
+  "@context": [
+    "https://www.w3.org/ns/credentials/v2",
+    "https://originator-profile.org/ns/credentials/v1",
+    "https://originator-profile.org/ns/cip/v1",
+    {
+      "@language": "ja-JP"
+    }
+  ],
+  "type": ["VerifiableCredential", "ProfileAnnotation"],
+  "issuer": "dns:localhost",
+  "credentialSubject": {
+    "id": "dns:localhost",
+    "type": "JP-OrganizationExistenceCertificate",
+    "corporateName": "開発用サンプル組織",
+    "corporateNumber": "0000000000000",
+    "postalCode": "000-0000",
+    "addressCountry": "JP",
+    "addressRegion": "東京都",
+    "addressLocality": "千代田区",
+    "streetAddress": "○○○",
+    "annotation": {
+      "id": "urn:uuid:def09cbd-6e8e-4c73-856d-5e00dffde643",
+      "type": "ProfileAnnotationPolicy",
+      "name": "架空組織実在性検証局 実在証明",
+      "description": "この組織は、法人登記の照会等により組織が実在していることが確認できました。",
+      "ref": "https://ovac.exp.originator-profile.org/"
+    }
+  }
+}
