+++ packages/verify/src/originator-profile-set/media.ts
@@ -1,0 +1,34 @@
+import { WebMediaProfile } from "@originator-profile/model";
+import {
+  UnverifiedJwtVc,
+  VcValidator,
+} from "@originator-profile/securing-mechanism";
+import { verifyImageDigestSri } from "../integrity";
+import { type MappedKeys } from "../keys";
+import { OpVerifier } from "./op-verifier";
+
+/** media プロパティの署名検証 */
+export async function verifyMedia(
+  wmpIssuerKeys: MappedKeys,
+  media?: UnverifiedJwtVc<WebMediaProfile>[],
+  validator?: typeof VcValidator,
+) {
+  if (!media) return;
+  return await Promise.all(
+    media.map(async (m) => {
+      const verify = OpVerifier<WebMediaProfile>(
+        wmpIssuerKeys,
+        m,
+        validator?.(WebMediaProfile),
+      );
+      const result = await verify(m.source);
+      if (result instanceof Error) {
+        return result;
+      }
+
+      await verifyImageDigestSri(result.doc.credentialSubject.logo);
+
+      return result;
+    }),
+  );
+}
